Make a single Kubernetes role
This commit puts all the different roles (common, master and worker)
into a single one.
It also sets just one group for kubernetes hosts named kubernetes_nodes
and differentiate the role of each node with the kubernetes_role
variable, which can be 'master' or 'worker'.
It finally renames the main inventory from the specific "graylog-lab" to
the general "lab".