Make Azure provider variables hidden
This commit takes away from the produced terraform files the relevant
information related to the login (so subscription_id, client_id,
client_secret and tenant_id) and put them in a source file, kept in the
azure-private directory, meant to be stored away from the main terraform
direcotory.
It also fixes the ssh keys creation so that if the playbook need to
create the keys, those get stored into the same azure-private folder
created for the source files described above.