Rename template into vms for terraform
This commit renames the general template.tf.j2 that was related to the
virtual machines .tf files generation into vms.tf.j2 which is more self
explaining. The result is that into the azure tf directory there will be
a sequence of vms_<name>.yml file for each group of virtual machines.