Fix Python requirements for Azure
This commit fixes the Python requirements and reverts the dependabot
patch over the cryptography component suggested by GitHub in october.
Since the 3.2 release (and the entire 3.0 releases series) is not (yet)
compatible with azure-cli, if we still want to use it we need to lower
down the cryptography version back to 2.9.2.