Support private endpoints with Terraform Azure DBs
This commit adds the support for creating Azure's private enpoints
while creating Azure's databases.
It is sufficient to add the relative section in the configuration:

azure_database_servers:
  - name: 'mymasterdb'
    kind: 'MariaDB'
    version: '10.2'
    ...
    ...
    # Private endpoints
    private_endpoints:
      - name: 'mydbep'
        subnet_id: 'my_subnet_server'

Since it is required to make private endpoints work, an additional
param named enforce_private_link_endpoint_network_policies has been
added inside the subnets declarations, so that it is possible to set
it true while creating the associated private endpoint.
Like this:

azure_vnets:
  - name: 'ausdev_vnet_01'
    addr_space: '192.168.118.0/23'
    ...
    ...
    subnets:
      - name: 'my_subnet_server'
        prefixes:
          - '192.168.1.0/24'
        enforce_private_link_endpoint_network_policies: 'true'